Move even more metrics to `vm_interfaces`
