+++ core/lib/dal/src/transactions_web3_dal.rs
@@ -479,9 +479,8 @@ impl TransactionsWeb3Dal<'_, '_> {
 mod tests {
     use std::collections::HashMap;
 
-    use zksync_types::{
-        fee::TransactionExecutionMetrics, l2::L2Tx, Nonce, ProtocolVersion, ProtocolVersionId,
-    };
+    use zksync_types::{l2::L2Tx, Nonce, ProtocolVersion, ProtocolVersionId};
+    use zksync_vm_interface::TransactionExecutionMetrics;
 
     use super::*;
     use crate::{
