+++ core/lib/multivm/src/versions/vm_latest/implementation/statistics.rs
@@ -1,8 +1,8 @@
 use zk_evm_1_5_0::aux_structures::Timestamp;
-use zksync_types::{circuit::CircuitStatistic, U256};
+use zksync_types::U256;
 
 use crate::{
-    interface::{storage::WriteStorage, VmExecutionStatistics, VmMemoryMetrics},
+    interface::{storage::WriteStorage, CircuitStatistic, VmExecutionStatistics, VmMemoryMetrics},
     vm_latest::vm::Vm,
     HistoryMode,
 };
