+++ core/lib/vm_interface/src/lib.rs
@@ -8,10 +8,11 @@ pub use crate::{
         },
         inputs::{L1BatchEnv, L2BlockEnv, SystemEnv, TxExecutionMode, VmExecutionMode},
         outputs::{
-            BootloaderMemory, CompressedBytecodeInfo, CurrentExecutionState,
+            BootloaderMemory, CircuitStatistic, CompressedBytecodeInfo, CurrentExecutionState,
             DeduplicatedWritesMetrics, ExecutionResult, FinishedL1Batch, L2Block, Refunds,
-            TransactionExecutionResult, TxExecutionStatus, VmExecutionLogs, VmExecutionMetrics,
-            VmExecutionResultAndLogs, VmExecutionStatistics, VmMemoryMetrics,
+            TransactionExecutionMetrics, TransactionExecutionResult, TxExecutionStatus,
+            VmExecutionLogs, VmExecutionMetrics, VmExecutionResultAndLogs, VmExecutionStatistics,
+            VmMemoryMetrics,
         },
         tracer,
     },
