+++ core/node/api_server/src/execution_sandbox/execute.rs
@@ -4,13 +4,15 @@ use anyhow::Context as _;
 use tracing::{span, Level};
 use zksync_dal::{ConnectionPool, Core};
 use zksync_multivm::{
-    interface::{TxExecutionMode, VmExecutionResultAndLogs, VmInterface},
+    interface::{
+        TransactionExecutionMetrics, TxExecutionMode, VmExecutionResultAndLogs, VmInterface,
+    },
     tracers::StorageInvocations,
     MultiVMTracer,
 };
 use zksync_types::{
-    fee::TransactionExecutionMetrics, l2::L2Tx, transaction_request::CallOverrides,
-    ExecuteTransactionCommon, Nonce, PackedEthSignature, Transaction, U256,
+    l2::L2Tx, transaction_request::CallOverrides, ExecuteTransactionCommon, Nonce,
+    PackedEthSignature, Transaction, U256,
 };
 
 use super::{
