+++ core/node/state_keeper/src/io/tests/mod.rs
@@ -4,12 +4,14 @@ use test_casing::test_casing;
 use zksync_contracts::BaseSystemContractsHashes;
 use zksync_dal::{ConnectionPool, Core, CoreDal};
 use zksync_mempool::L2TxFilter;
-use zksync_multivm::{interface::VmExecutionMetrics, utils::derive_base_fee_and_gas_per_pubdata};
+use zksync_multivm::{
+    interface::{TransactionExecutionMetrics, VmExecutionMetrics},
+    utils::derive_base_fee_and_gas_per_pubdata,
+};
 use zksync_node_test_utils::prepare_recovery_snapshot;
 use zksync_types::{
     block::{BlockGasCount, L2BlockHasher},
     commitment::L1BatchCommitmentMode,
-    fee::TransactionExecutionMetrics,
     fee_model::{BatchFeeInput, PubdataIndependentBatchFeeModelInput},
     AccountTreeId, Address, L1BatchNumber, L2BlockNumber, L2ChainId, ProtocolVersion,
     ProtocolVersionId, StorageKey, VmEvent, H256, U256,
