+++ core/node/state_keeper/src/io/tests/tester.rs
@@ -11,7 +11,8 @@ use zksync_contracts::BaseSystemContracts;
 use zksync_dal::{ConnectionPool, Core, CoreDal};
 use zksync_eth_client::{clients::MockSettlementLayer, BaseFees};
 use zksync_multivm::{
-    interface::TransactionExecutionResult, vm_latest::constants::BATCH_COMPUTATIONAL_GAS_LIMIT,
+    interface::{TransactionExecutionMetrics, TransactionExecutionResult},
+    vm_latest::constants::BATCH_COMPUTATIONAL_GAS_LIMIT,
 };
 use zksync_node_fee_model::{
     l1_gas_price::{GasAdjuster, GasAdjusterClient},
@@ -24,7 +25,6 @@ use zksync_node_test_utils::{
 use zksync_types::{
     block::L2BlockHeader,
     commitment::L1BatchCommitmentMode,
-    fee::TransactionExecutionMetrics,
     fee_model::{BatchFeeInput, FeeModelConfig, FeeModelConfigV1},
     l2::L2Tx,
     protocol_version::{L1VerifierConfig, ProtocolSemanticVersion},
