+++ core/node/state_keeper/src/seal_criteria/criteria/geometry_seal_criteria.rs
@@ -69,8 +69,7 @@ impl SealCriterion for CircuitsCriterion {
 }
 #[cfg(test)]
 mod tests {
-    use zksync_multivm::interface::VmExecutionMetrics;
-    use zksync_types::circuit::CircuitStatistic;
+    use zksync_multivm::interface::{CircuitStatistic, VmExecutionMetrics};
 
     use super::*;
